Game-server: переход на upstream 2.6.3-SNAPSHOT develop + порт PlayerPanel + фиксы
Полная замена прежней форк-версии l2j-server-game на свежий upstream L2J
(ветка develop с последними фиксами: transformations data loss, Eilhalder
respawn, siegeable CH exp restoration, Fear/Turn Undead, Multilingual,
Quest NPE) и апгрейд на JDK 25. Сохранены все кастомные модули и поверх
добавлены дополнительные security/bug-правки.

=== UPSTREAM ===
- Java 21 -> 25 (target в pom.xml).
- Event API переименован: OnXxx (класс/интерфейс) -> Xxx (record extends BaseEvent).
- slf4j 2.0.9 -> 2.0.17, log4j 2.22.0 -> 2.25.3, gson 2.10.1 -> 2.13.2,
  jda 5.0.0-beta.18 -> 5.3.0, l2j-server-commons 2.6.7.0 -> 2.6.9.0,
  junit 5.10.1 -> 6.0.1, mockito 5.7.0 -> 5.21.0, assertj 3.24.2 -> 3.27.6,
  maven-compiler 3.11.0 -> 3.14.1, surefire 3.2.2 -> 3.5.4.
- Удалены мёртвые old-API классы: TeleportLocationTable, MerchantPriceConfigTable,
  L2AdventurerInstance, L2TrainerInstance, IBaseEvent и набор OnXxx events.

=== ПОРТ КАСТОМНЫХ МОДУЛЕЙ ===
PlayerPanel целиком перенесён и адаптирован под upstream API:
- bbs/service/PlayerPanelBoardHandler + modules/* (Shop, Enchant, Buff,
  Validator, HtmlGenerator, Utils).
- config/PlayerPanelConfig, LuckyPigConfig, AdvancedRewardsConfig,
  EnhancedRewardsOnlineConfig.
- dao/impl/mysql/PlayerPanelDAO (in-memory кулдауны, async логирование).
- loader/PlayerPanelLoader (EventType.PLAYER_LOGOUT + PlayerLogout record).
- Интеграция: Configuration.PlayerPanel()/LuckyPig() + GameServer.java
  вызов PlayerPanelLoader.load() в BBS-секции.

=== SECURITY / BUG ФИКСЫ ПОВЕРХ UPSTREAM ===
Критичное:
- SecondaryPasswordAuth: SHA-1 без соли -> PBKDF2-SHA256 (120k iter) + lazy
  миграция со старого формата при первой успешной проверке.
- BitSetIDFactory: NullPointerException на исчерпании Id -> IllegalStateException.

Важное:
- ThreadPoolManager.shutdown(): правильный порядок (shutdown перед
  awaitTermination, было обратное), добавлены _aiScheduledThreadPool
  и _eventScheduledThreadPool (не останавливались совсем).
- LoginServerThread: 8 пустых catch(IOException){} заменены на LOG.warn
  в sendAccessLevel/sendClientTracert/sendMail/sendTempBan/sendServerStatus/
  sendServerType/sendChangePassword/getCharsOnServer.
- Shutdown.java: 5 пустых catch(Exception){} заменены на LOG.warn + dedicated
  COUNTER_LOCK вокруг _counterInstance для защиты параллельных /shutdown.
- IdFactory.cleanUpTimeStamps: пустой catch -> LOG.warn.
- TaskManager (TYPE_TIME): пустой catch -> LOG.warn с контекстом task.getId().
- StackIDFactory: java.util.Stack -> ArrayDeque (убран ненужный внутренний
  Vector-lock при внешнем synchronized).

Минор:
- L2World.getVisibleObjects/3D/Playable: LinkedList -> ArrayList(pre-sized)
  в hot-path AI-тиков.
- GameTimeController.run(): InterruptedException теперь сохраняет
  interrupt-flag и выходит из цикла.
- FloodProtectorAction.log: пустой catch при IP-lookup -> LOG.trace.

PlayerPanel polish:
- Shop: цены HTML синхронизированы с серверным CATALOG (публичный static
  final Map + ShopEntry public), убрано расхождение UX.
- HTML currency: убран copy-paste дубликат Fame для фейковых L2-Coin/Honor,
  оставлены Adena/Fame/PvP.
- BuffModule.applyBuffScheme: refund при appliedCount == 0 (ранее игрок
  мог заплатить и получить 0 бафов).
- Все reduceAdena/addAdena в BuffModule теперь передают player в reference
  вместо null (audit-trail).
- PlayerPanelConfig/AdvancedRewardsConfig/EnhancedRewardsOnlineConfig:
  @Sources теперь имеет 3 источника (${L2J_HOME}/custom/game/config/,
  ./config/, classpath) по паттерну LuckyPigConfig.

Сборка: JDK 25, mvnw package проходит чисто (1829 файлов).

+++ l2j-server-game/src/main/java/com/l2jserver/gameserver/config/GraciaSeedsConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2004-2023 L2J Server
+ * Copyright © 2004-2026 L2J Server
  *
  * This file is part of L2J Server.
  *
@@ -44,9 +44,9 @@ import com.l2jserver.gameserver.config.converter.Minutes2MillisecondsConverter;
 public interface GraciaSeedsConfiguration extends Reloadable {
 	
 	@Key("TiatKillCountForNextState")
-	Integer getTiatKillCountForNextState();
+	int getTiatKillCountForNextState();
 	
 	@Key("Stage2Length")
 	@ConverterClass(Minutes2MillisecondsConverter.class)
-	Integer getStage2Length();
+	long getStage2Length();
 }
